debug:法器符术的文本显示问题

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/runes/spells/BlessSpell.java
@@ -26,7 +26,7 @@ public class BlessSpell extends TargetedSpell{
     }
 
     public String desc(){
-        float power = 1;
+        int power = 1;
         return Messages.get(this, "desc", 2+4*power, 5+5*power, 5+5*power, 5+5*power) +
                 "\n\n" + Type() + Messages.get(this, "overrunes", (int)overRunes(Dungeon.hero));
     }
